# api_clients/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"DefaultSelection_Prisma._36_ProfessorPayload_": {
				"properties": {
					"school_id": {
						"type": "integer",
						"format": "int32"
					},
					"rmp_id": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"first_name": {
						"type": "string"
					},
					"id": {
						"type": "integer",
						"format": "int32"
					}
				},
				"required": [
					"school_id",
					"rmp_id",
					"last_name",
					"first_name",
					"id"
				],
				"type": "object"
			},
			"Professor": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_ProfessorPayload_",
				"description": "Model Professor"
			},
			"ProfessorSearchResult": {
				"properties": {
					"pageInfo": {
						"properties": {
							"total": {
								"type": "integer",
								"format": "int32"
							},
							"endCursor": {
								"type": "string",
								"nullable": true
							},
							"hasNextPage": {
								"type": "boolean"
							}
						},
						"required": [
							"total",
							"endCursor",
							"hasNextPage"
						],
						"type": "object"
					},
					"edges": {
						"items": {
							"properties": {
								"node": {
									"$ref": "#/components/schemas/Professor"
								},
								"cursor": {
									"type": "string"
								}
							},
							"required": [
								"node",
								"cursor"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"pageInfo",
					"edges"
				],
				"type": "object"
			},
			"ProfessorErrorType": {
				"enum": [
					"PROFESSOR_NOT_FOUND",
					"PROFESSOR_ALREADY_EXISTS"
				],
				"type": "string"
			},
			"Pick_ProfessorError.message-or-type_": {
				"properties": {
					"message": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"properties": {},
								"additionalProperties": {
									"type": "string"
								},
								"type": "object"
							}
						]
					},
					"type": {
						"$ref": "#/components/schemas/ProfessorErrorType"
					}
				},
				"required": [
					"message",
					"type"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"ProfessorErrorType.PROFESSOR_ALREADY_EXISTS": {
				"enum": [
					"PROFESSOR_ALREADY_EXISTS"
				],
				"type": "string"
			},
			"ProfessorAlreadyExistsError": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Pick_ProfessorError.message-or-type_"
					},
					{
						"properties": {
							"type": {
								"$ref": "#/components/schemas/ProfessorErrorType.PROFESSOR_ALREADY_EXISTS"
							}
						},
						"required": [
							"type"
						],
						"type": "object"
					}
				]
			},
			"Pick_Professor.first_name-or-last_name-or-rmp_id-or-school_id_": {
				"properties": {
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"rmp_id": {
						"type": "string"
					},
					"school_id": {
						"type": "integer",
						"format": "int32"
					}
				},
				"required": [
					"first_name",
					"last_name",
					"rmp_id",
					"school_id"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"NewProfessor": {
				"$ref": "#/components/schemas/Pick_Professor.first_name-or-last_name-or-rmp_id-or-school_id_"
			},
...
			"ReviewTag": {
				"$ref": "#/components/schemas/_36_Enums.ReviewTag"
			},
			"TagAmount": {
				"properties": {
					"amount": {
						"type": "integer",
						"format": "int32"
					},
					"tag": {
						"$ref": "#/components/schemas/ReviewTag"
					}
				},
				"required": [
					"amount",
					"tag"
				],
				"type": "object"
			},
			"ChartValue": {
				"properties": {
					"year": {
						"type": "integer",
						"format": "int32"
					},
					"month": {
						"type": "string"
					},
					"value": {
						"type": "integer",
						"format": "int32"
					}
				},
				"required": [
					"year",
					"month",
					"value"
				],
				"type": "object"
			},
			"ProfessorAnalysis": {
				"properties": {
					"averageRatingValues": {
						"items": {
							"$ref": "#/components/schemas/ChartValue"
						},
						"type": "array"
					},
					"tagAmount": {
						"items": {
							"$ref": "#/components/schemas/TagAmount"
						},
						"type": "array"
					}
				},
				"required": [
					"averageRatingValues",
					"tagAmount"
				],
				"type": "object"
			},
			"ProfessorErrorType.PROFESSOR_NOT_FOUND": {
				"enum": [
					"PROFESSOR_NOT_FOUND"
				],
				"type": "string"
			},
			"ProfessorNotFoundError": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Pick_ProfessorError.message-or-type_"
					},
					{
						"properties": {
							"type": {
								"$ref": "#/components/schemas/ProfessorErrorType.PROFESSOR_NOT_FOUND"
							}
						},
						"required": [
							"type"
						],
						"type": "object"
					}
				]
			},
...
			"Grade": {
				"$ref": "#/components/schemas/_36_Enums.Grade"
			},
			"Rating": {
				"properties": {
					"averageGrade": {
						"$ref": "#/components/schemas/Grade"
					},
					"topKMostRecentDifficultyAverage": {
						"type": "number",
						"format": "float"
					},
					"totalDifficultyAverage": {
						"type": "number",
						"format": "float"
					},
					"topKMostRecentQualityAverage": {
						"type": "number",
						"format": "float"
					},
					"totalQualityAverage": {
						"type": "number",
						"format": "float"
					},
					"ratingAmount": {
						"type": "integer",
						"format": "int32"
					}
				},
				"required": [
					"averageGrade",
					"topKMostRecentDifficultyAverage",
					"totalDifficultyAverage",
					"topKMostRecentQualityAverage",
					"totalQualityAverage",
					"ratingAmount"
				],
				"type": "object"
			},
...
			"Course": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_CoursePayload_",
				"description": "Model Course"
			},
			"_36_Enums.Semester": {
				"type": "string",
				"enum": [
					"SPRING",
					"SUMMER",
					"FALL"
				]
			},
			"Semester": {
				"$ref": "#/components/schemas/_36_Enums.Semester"
			},
			"ProfessorCourses": {
				"properties": {
					"total": {
						"type": "integer",
						"format": "int32"
					},
					"courses": {
						"items": {
							"allOf": [
								{
									"$ref": "#/components/schemas/Course"
								},
								{
									"properties": {
										"year": {
											"type": "integer",
											"format": "int32"
										},
										"semester": {
											"$ref": "#/components/schemas/Semester"
										}
									},
									"required": [
										"year",
										"semester"
									],
									"type": "object"
								}
							]
						},
						"type": "array"
					}
				},
				"required": [
					"total",
					"courses"
				],
				"type": "object"
			},
			"UpdatedProfessor": {
				"properties": {
					"last_name": {
						"type": "string",
						"nullable": true
					},
					"first_name": {
						"type": "string",
						"nullable": true
					}
				},
				"type": "object"
			},
			"ReviewResponse": {
				"properties": {
					"id": {
						"type": "integer",
						"format": "int32"
					},
					"quality": {
						"type": "number",
						"format": "float"
					},
					"difficulty": {
						"type": "number",
						"format": "float"
					},
					"time": {
						"type": "string",
						"format": "date-time"
					},
					"tags": {
						"items": {
							"$ref": "#/components/schemas/ReviewTag"
						},
						"type": "array"
					},
					"grade": {
						"$ref": "#/components/schemas/Grade"
					},
					"professor_id": {
						"type": "integer",
						"format": "int32"
					}
				},
				"required": [
					"id",
					"quality",
					"difficulty",
					"time",
					"tags",
					"grade",
					"professor_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewsSearchResult": {
				"properties": {
					"pageInfo": {
						"properties": {
							"total": {
								"type": "integer",
								"format": "int32"
							},
							"endCursor": {
								"type": "string",
								"nullable": true
							},
							"hasNextPage": {
								"type": "boolean"
							}
						},
						"required": [
							"total",
							"endCursor",
							"hasNextPage"
						],
						"type": "object"
					},
					"edges": {
						"items": {
							"properties": {
								"node": {
									"$ref": "#/components/schemas/ReviewResponse"
								},
								"cursor": {
									"type": "string"
								}
							},
							"required": [
								"node",
								"cursor"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"pageInfo",
					"edges"
				],
				"type": "object"
			},
...
			"ProfessorCourse": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_ProfessorCoursePayload_",
				"description": "Model ProfessorCourse"
			},
			"CourseEnrollmentResult": {
				"$ref": "#/components/schemas/ProfessorCourse"
			},
...
			"Omit_ProfessorCourse.professor_id-or-id_": {
				"$ref": "#/components/schemas/Pick_ProfessorCourse.Exclude_keyofProfessorCourse.professor_id-or-id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"CourseEnrollment": {
				"$ref": "#/components/schemas/Omit_ProfessorCourse.professor_id-or-id_"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/professors/search": {
			"get": {
				"operationId": "SearchProfessors",
				"responses": {
					"200": {
						"description": "Professors Retrieved Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfessorSearchResult"
								}
							}
						}
					}
				},
				"tags": [
					"Professors"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "school_id",
						"required": false,
						"schema": {
							"format": "int32",
							"type": "integer"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "int32",
							"type": "integer"
						}
					}
				]
			}
		},
		"/professors/create": {
			"post": {
				"operationId": "CreateProfessor",
				"responses": {
					"201": {
						"description": "Professor Created Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Professor"
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": 1,
											"first_name": "Example",
											"last_name": "Professor",
											"rmp_id": "hfaksjdfhuicyv",
											"school_id": 1
										}
									}
								}
							}
						}
					},
					"409": {
						"description": "Professor already exists",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfessorAlreadyExistsError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Professors"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NewProfessor"
							}
						}
					}
				}
			}
		},
		"/professors/{id}/analysis": {
			"get": {
				"operationId": "GetProfessorAnalysis",
				"responses": {
					"200": {
						"description": "Professor Analysis Retrieved Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfessorAnalysis"
								}
							}
						}
					},
					"404": {
						"description": "Professor not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfessorNotFoundError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Professors"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer"
						}
					}
				]
			}
		},
		"/professors/{id}/rating": {
			"get": {
				"operationId": "GetProfessorRating",
				"responses": {
					"200": {
						"description": "Rating Retrieved Successfully",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Rating"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Professor not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfessorNotFoundError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Professors"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer"
						}
					},
					{
						"in": "query",
						"name": "topKPercentage",
						"required": false,
						"schema": {
							"format": "int32",
							"type": "integer"
						}
					}
				]
			}
		},
		"/professors/{id}": {
			"get": {
				"operationId": "GetProfessor",
				"responses": {
					"200": {
						"description": "Professor Retrieved Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Professor"
								}
							}
						}
					},
					"404": {
						"description": "Professor not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfessorNotFoundError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Professors"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateProfessor",
				"responses": {
					"200": {
						"description": "Professor Updated Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Professor"
								}
							}
						}
					},
					"404": {
						"description": "Professor not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfessorNotFoundError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Professors"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatedProfessor"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProfessor",
				"responses": {
					"200": {
						"description": "Professor Deleted Successfully"
					},
					"404": {
						"description": "Professor not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfessorNotFoundError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Professors"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer"
						}
					}
				]
			}
		},
		"/professors/{id}/courses": {
			"get": {
				"operationId": "GetProfessorCourses",
				"responses": {
					"200": {
						"description": "Courses Retrieved Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfessorCourses"
								}
							}
						}
					},
					"404": {
						"description": "Professor not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfessorNotFoundError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Professors"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer"
						}
					},
					{
						"in": "query",
						"name": "year",
						"required": false,
						"schema": {
							"format": "int32",
							"type": "integer"
						}
					},
					{
						"in": "query",
						"name": "semester",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/Semester"
						}
					}
				]
			}
		},
		"/professors/{id}/reviews": {
			"get": {
				"operationId": "GetReviews",
				"responses": {
					"200": {
						"description": "Reviews Retrieved Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ReviewsSearchResult"
								}
							}
						}
					}
				},
				"tags": [
					"Professors"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 25,
							"format": "int32",
							"type": "integer"
						}
					}
				]
			}
		},
		"/professors/{id}/enroll": {
			"post": {
				"operationId": "Enroll",
				"responses": {
					"201": {
						"description": "Review Created Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CourseEnrollmentResult"
								}
							}
						}
					}
				},
				"tags": [
					"Professors"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CourseEnrollment"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}